+++ tests/e2e/tests/editors/object-enforceAdditionalProperties_test.cjs
@@ -42,3 +42,20 @@ Scenario('@editor @object @enforceAdditionalProperties should remove all propert
     "id_": "test"
   }))
 })
+
+Scenario('@editor @object @enforceAdditionalProperties `Add property` should only accept a name matching `properties` or `patternProperties`', async ({ I }) => {
+  I.amOnPage(`playground.html?theme=${theme}`)
+  I.selectOption('#examples', 'editors/object-enforceAdditionalProperties')
+  I._waitForElement('.jedi-ready')
+
+  I._click('.jedi-quick-add-property-toggle')
+  I.fillField('#jedi-quick-add-property-input-root', 'id_test')
+  I._click('.jedi-add-property-btn')
+  I._waitForElement('[data-path="#/id_test"]')
+
+  I._click('.jedi-quick-add-property-toggle')
+  I.fillField('#jedi-quick-add-property-input-root', 'c')
+  I._click('.jedi-add-property-btn')
+  I._waitForElement('#jedi-quick-add-property-input-root-messages .jedi-error-message')
+  I.dontSeeElement('[data-path="#/c"]')
+})